Abort FDN delete when number is empty (instead of name)
Changed FDN delete to abort when number is empty rather than
when name is empty

3GPP TS 31.102 (USIM application) v9.1.0 states in chapter
4.2.24 EFFDN (Fixed Dialling Number) that the Dialling Number
field is mandatory whereas the Alpha Identifier is optional.
The current implementation of the user interface for editing
the fixed dialling list blocks creation and deletion list
entries without a name (corresponding to the Alpha Identifier)
but instead permits entries without a number. This is incompatible
with the USIM specification and
implementation.

This contribution corrects the implementation in line with
the USIM specification.

Change-Id: I7e2ca06e20401b517d705fc45c8b873b5c21cc4c