Add test runner log when test case class is not loaded.
The test runner will only load TestCase classes that have a public
constructor with either no params or a single String parameter.

This is fairly subtle behavior and can be difficult to catch. This change
adds a log message when a TestCase class is not loaded, and adds some
associated unit tests.

Change-Id: I6fc63e4179c949620f5773e0ae134f99905a6fb2