Frameworks changes to enable File Reader and blob.slice APIs.
Rather than just storing the name of the file that the user chose in the file picker,
store the full path to the file. That way we can open it using the file API and still
extract the filename itself by simply chopping off all but the last path segment.

Change-Id: I25f2fe818deda8ca848f0da02cf45ae97d2127fb