Move WebViewDatabase initialization into a background thread.
This isn't the perfect solution as any calls into WebViewDatabase will block
until initialization completes.  It is better than the current model which
blocks when creating the WebViewDatabase.

In order to get a perfect solution, we would need to do initialization in the
background and have a way to interact with the database during initialization
that would not block.

Bug: 2721802
Change-Id: Ibbaf7f97926128de6534fbf2148452099766c6b7