Use fixed viewpport for tablet device.
This makes the zoom animation much more smooth.  With previous dynamic
viewport, the animation jumps because of viewport adjustment and big
screen.

Also the viewport width is changed from 800 to 1040 to better suit
the tablet device.  It's the middle point of 1280 and 800.

The text reflow re-alignment for non-double-tap zoom is disabled for
tablet, since most of the time the reflow has no effect.

issue: 2919307
Change-Id: I74493df6408291c24ef8511512d7ed0b81bf507a