DO NOT MERGE Refactor find and select dialogs
(Merged by hand from master, based on
https://android-git.corp.google.com/g/#change,53490 )

In WebView:
- Remove the select mode, so that while text is selected,
the view may respond to scroll or zoom gestures.
- Remove the getFindIsUp() method -- let the host application
track if the dialog is visible.
- Map long press to select text if the host application does
not already intercept it.
- Draw find 'extras' during zoom animation.
- Draw select 'extras' during animation and scrolling as well.
- Distinguish between selecting text and pressing a shift key
(the shift key is still a possible select text shortcut)
- Add more public (but hidden) interfaces: setUpSelect, selectAll, etc.

In WebChromeClient, add interfaces to allow WebView to tell the host
application that selection has started or finished. A long press on
a part of the view that does not trigger a host context menu may
start the selection (if text is underneath the press), and clicking
the trackball or tapping away from the current selection may
end the selection.

Companion changes in external/webkit and packages/apps/Browser

Change-Id: I2d41662acbdb8bdaa8219b35f18a9276e34636a7
http://b/262451