Memory leak in TextKeyListener.
If a call to getPrefs is made after weak reference to content resolver
is gone, a new observer will be created and registered at the
resolver. At registration, the resolver will allocate ObserverEntry
objects that will eventually fill the JNI global reference pool.
There is no need to create and register a new observer if one is already
registered.

Change-Id: If8442b3370299980b73d4ea83757c5eae6a85408