Adds a test case for WebAddress where the path component does not have a leading slash
When a URL is malformed because the path does not start with a slash,
we assume that the path starts with the first character that is not
valid in the host and insert a leading slash. This is the reason why
the regex for the path component does not force a leading slash.

Bug: 1011602
Change-Id: I8efe46c058d2ee2d1a6a4406ee25dc021315222b