Undo API patch due to field pulled up to superclass.
Earlier we extracted a common superclass InputEvent from KeyEvent
and MotionEvent.  During this process, getDeviceId() and mDeviceId
were moved to the superclass.  The API change tracking was unable
to automatically figure out that the method had just been moved up
to the superclass (still binary compatible) so we patched up 9.xml
manually to fool it.  Unfortunately mDeviceId slipped in there
when it should not have since it should have been marked with @hide.
This change just removes mDeviceId.

Change-Id: I5fca14c8dc6f98952b8afbdb6e034824a22bdb1d